Specify min and max shiller from files
The market is reaching new highs even though we don't seem
to have had a real recession. So adding the ability to override values
more easily since it seems we hit new highs every other day.

Didn't need to do it for MINs yet but went ahead and did anyway. It does
make me stop and consider [1] and how to interpret it.

[1] https://en.wikipedia.org/wiki/List_of_recessions_in_the_United_States#:~:text=The%20NBER%20defines%20a%20recession,and%20wholesale%2Dretail%20sales%22.